Module msg::compositor_msg

Structs

Epoch
A newtype struct for denoting the age of messages; prevents race conditions.

LayerId

LayerMetadata
All layer-specific information that the painting task sends to the compositor other than the buffer contents of the layer itself.

Enums

PaintState
The status of the painter.

ReadyState

ScrollPolicy
The scrolling policy of a layer.

Traits

PaintListener
The interface used by the painter to acquire draw targets for each paint frame and submit them to be drawn to the display.

ScriptListener
The interface used by the script task to tell the compositor to update its ready state, which is used in displaying the appropriate message in the window's title.