Struct msg::compositor_msg::LayerMetadata

pub struct LayerMetadata {
    pub id: LayerId,
    pub position: Rect<i32>,
    pub background_color: Color,
    pub scroll_policy: ScrollPolicy,
}

All layer-specific information that the painting task sends to the compositor other than the buffer contents of the layer itself.

Fields

id

An opaque ID. This is usually the address of the flow and index of the box within it.

position

The position and size of the layer in pixels.

background_color

The background color of the layer.

scroll_policy

The scrolling policy of this layer.