Trait msg::compositor_msg::PaintListener

pub trait PaintListener {
    fn get_graphics_metadata(&mut self) -> Option<NativeGraphicsMetadata>;
    fn initialize_layers_for_pipeline(&mut self, pipeline_id: PipelineId, metadata: Vec<LayerMetadata>, epoch: Epoch);
    fn assign_painted_buffers(&mut self, pipeline_id: PipelineId, epoch: Epoch, replies: Vec<(LayerId, Box<LayerBufferSet>)>);
    fn paint_msg_discarded(&mut self);
    fn set_paint_state(&mut self, PipelineId, PaintState);
}

The interface used by the painter to acquire draw targets for each paint frame and submit them to be drawn to the display.

Required Methods

fn get_graphics_metadata(&mut self) -> Option<NativeGraphicsMetadata>

fn initialize_layers_for_pipeline(&mut self, pipeline_id: PipelineId, metadata: Vec<LayerMetadata>, epoch: Epoch)

Informs the compositor of the layers for the given pipeline. The compositor responds by creating and/or destroying paint layers as necessary.

fn assign_painted_buffers(&mut self, pipeline_id: PipelineId, epoch: Epoch, replies: Vec<(LayerId, Box<LayerBufferSet>)>)

Sends new buffers for the given layers to the compositor.

fn paint_msg_discarded(&mut self)

fn set_paint_state(&mut self, PipelineId, PaintState)