Trait msg::compositor_msg::ScriptListener

pub trait ScriptListener {
    fn set_ready_state(&mut self, PipelineId, ReadyState);
    fn scroll_fragment_point(&mut self, pipeline_id: PipelineId, layer_id: LayerId, point: Point2D<f32>);
    fn set_title(&mut self, pipeline_id: PipelineId, new_title: Option<String>);
    fn close(&mut self);
    fn dup(&mut self) -> Box<ScriptListener + 'static>;
    fn send_key_event(&mut self, key: Key, state: KeyState, modifiers: KeyModifiers);
}

The interface used by the script task to tell the compositor to update its ready state, which is used in displaying the appropriate message in the window's title.

Required Methods

fn set_ready_state(&mut self, PipelineId, ReadyState)

fn scroll_fragment_point(&mut self, pipeline_id: PipelineId, layer_id: LayerId, point: Point2D<f32>)

fn set_title(&mut self, pipeline_id: PipelineId, new_title: Option<String>)

Informs the compositor that the title of the page with the given pipeline ID has changed.

fn close(&mut self)

fn dup(&mut self) -> Box<ScriptListener + 'static>

fn send_key_event(&mut self, key: Key, state: KeyState, modifiers: KeyModifiers)