Enum msg::constellation_msg::Msg

pub enum Msg {
    Exit,
    Failure(Failure),
    InitLoadUrl(Url),
    LoadComplete,
    FrameRect(PipelineId, SubpageId, Rect<f32>),
    LoadUrl(PipelineId, LoadData),
    ScriptLoadedURLInIFrame(Url, PipelineId, SubpageId, Option<SubpageId>, IFrameSandboxState),
    Navigate(Option<(PipelineId, SubpageId)>, NavigationDirection),
    PainterReady(PipelineId),
    ResizedWindow(WindowSizeData),
    KeyEvent(Key, KeyState, KeyModifiers),
    GetPipelineTitle(PipelineId),
    SetCursor(Cursor),
    MozBrowserEventMsg(PipelineId, SubpageId, MozBrowserEvent),
    ChangeRunningAnimationsState(PipelineId, bool),
    TickAnimation(PipelineId),
}

Messages from the compositor and script to the constellation.

Variants

Exit

Failure

InitLoadUrl

LoadComplete

FrameRect

LoadUrl

ScriptLoadedURLInIFrame

Navigate

PainterReady

ResizedWindow

KeyEvent

GetPipelineTitle

Requests that the constellation inform the compositor of the title of the pipeline immediately.

SetCursor

Requests that the constellation inform the compositor of the a cursor change.

MozBrowserEventMsg

Dispatch a mozbrowser event to a given iframe. Only available in experimental mode.

ChangeRunningAnimationsState

Indicates whether this pipeline is currently running animations.

TickAnimation

Requests that the constellation instruct layout to begin a new tick of the animation.