Struct msg::constellation_msg::WindowSizeData

pub struct WindowSizeData {
    pub initial_viewport: TypedSize2D<ViewportPx, f32>,
    pub visible_viewport: TypedSize2D<PagePx, f32>,
    pub device_pixel_ratio: ScaleFactor<ViewportPx, DevicePixel, f32>,
}

Fields

initial_viewport

The size of the initial layout viewport, before parsing an http://www.w3.org/TR/css-device-adapt/#initial-viewport

visible_viewport

The "viewing area" in page px. See PagePx documentation for details.

device_pixel_ratio

The resolution of the window in dppx, not including any "pinch zoom" factor.