Type Alias msg::constellation_msg::WebViewId

pub type WebViewId = TopLevelBrowsingContextId;

Aliased Type

struct WebViewId(pub BrowsingContextId);

Fields
0: BrowsingContextId
Implementations

impl TopLevelBrowsingContextId

pub fn new() -> TopLevelBrowsingContextId

pub fn install(id: TopLevelBrowsingContextId)

Each script and layout thread should have the top-level browsing context id installed, since it is used by crash reporting.

pub fn installed() -> Option<TopLevelBrowsingContextId>