Module msg::constellation_msg

The high-level interface from script to constellation. Using this abstract interface helps reduce coupling between these two components.

Structs

ConstellationChan

FrameId

Image

KeyModifiers

LoadData
Similar to net::resource_thread::LoadData can be passed to LoadUrl to load a page with GET/POST parameters or headers

PipelineId

PipelineIndex

PipelineNamespace
Each pipeline ID needs to be unique. However, it also needs to be possible to generate the pipeline ID from an iframe element (this simplifies a lot of other code that makes use of pipeline IDs).

PipelineNamespaceId

SubpageId

WindowSizeData

Enums

Key

KeyState

NavigationDirection

PixelFormat

ReferrerPolicy
Policies for providing a referrer header for a request

WebDriverCommandMsg

WindowSizeType

Constants

ALT

CONTROL

NONE

SHIFT

SUPER

_IMPL_DESERIALIZE_FOR_ConstellationChan

_IMPL_DESERIALIZE_FOR_FrameId

_IMPL_DESERIALIZE_FOR_Image

_IMPL_DESERIALIZE_FOR_Key

_IMPL_DESERIALIZE_FOR_KeyModifiers

_IMPL_DESERIALIZE_FOR_KeyState

_IMPL_DESERIALIZE_FOR_LoadData

_IMPL_DESERIALIZE_FOR_NavigationDirection

_IMPL_DESERIALIZE_FOR_PipelineId

_IMPL_DESERIALIZE_FOR_PipelineIndex

_IMPL_DESERIALIZE_FOR_PipelineNamespaceId

_IMPL_DESERIALIZE_FOR_PixelFormat

_IMPL_DESERIALIZE_FOR_ReferrerPolicy

_IMPL_DESERIALIZE_FOR_SubpageId

_IMPL_DESERIALIZE_FOR_WebDriverCommandMsg

_IMPL_DESERIALIZE_FOR_WindowSizeData

_IMPL_DESERIALIZE_FOR_WindowSizeType

_IMPL_SERIALIZE_FOR_ConstellationChan

_IMPL_SERIALIZE_FOR_FrameId

_IMPL_SERIALIZE_FOR_Image

_IMPL_SERIALIZE_FOR_Key

_IMPL_SERIALIZE_FOR_KeyModifiers

_IMPL_SERIALIZE_FOR_KeyState

_IMPL_SERIALIZE_FOR_LoadData

_IMPL_SERIALIZE_FOR_NavigationDirection

_IMPL_SERIALIZE_FOR_PipelineId

_IMPL_SERIALIZE_FOR_PipelineIndex

_IMPL_SERIALIZE_FOR_PipelineNamespaceId

_IMPL_SERIALIZE_FOR_PixelFormat

_IMPL_SERIALIZE_FOR_ReferrerPolicy

_IMPL_SERIALIZE_FOR_SubpageId

_IMPL_SERIALIZE_FOR_WebDriverCommandMsg

_IMPL_SERIALIZE_FOR_WindowSizeData

_IMPL_SERIALIZE_FOR_WindowSizeType

Statics

PIPELINE_NAMESPACE

Type Definitions

PanicMsg