Module msg::constellation_msg

The high-level interface from script to constellation. Using this abstract interface helps reduce coupling between these two components.

Structs

ConstellationChan

Failure

FrameId

KeyModifiers

LoadData
Similar to net::resource_task::LoadData can be passed to LoadUrl to load a page with GET/POST parameters or headers

PipelineId

SubpageId

WindowSizeData

WorkerId

Enums

IFrameSandboxState

Key

KeyState

MozBrowserEvent

Msg
Messages from the compositor and script to the constellation.

NavigationDirection

PipelineExitType

Constants

ALT

CONTROL

SHIFT

SUPER