Trait msg::constellation_msg::BackgroundHangMonitorRegister

pub trait BackgroundHangMonitorRegister: BackgroundHangMonitorClone + Send {
    fn register_component(
        &self, 
        component: MonitoredComponentId, 
        transient_hang_timeout: Duration, 
        permanent_hang_timeout: Duration, 
        exit_signal: Option<Box<dyn BackgroundHangMonitorExitSignal>>
    ) -> Box<dyn BackgroundHangMonitor>;
}

A handle to register components for hang monitoring, and to receive a means to communicate with the underlying hang monitor worker.

Required methods

fn register_component(
&self,
component: MonitoredComponentId,
transient_hang_timeout: Duration,
permanent_hang_timeout: Duration,
exit_signal: Option<Box<dyn BackgroundHangMonitorExitSignal>>
) -> Box<dyn BackgroundHangMonitor>

Register a component for hang monitoring: to be called from within the thread to be monitored for hangs.